+++ BrainflowCyton/__init__.py
@@ -4,5 +4,5 @@ BrainflowCyton.
 Python wrapper for BrainFlow API + Cyton SDK for collecting realtime or offline (SDCard) EEG, EMG, EKG data.
 """
 
-__version__ = "0.2.1"
+__version__ = "0.2.2"
 __author__ = 'zeyus'